+++ CHANGELOG.md
@@ -121,5 +121,10 @@ The first release of LambdaBetterGrass!
 - Fixed several issues with custom resource packs with custom grass models ([#48](https://github.com/LambdAurora/LambdaBetterGrass/issues/48) and partially [#76](https://github.com/LambdAurora/LambdaBetterGrass/issues/76)).
 - (1.20) Added better snow data to the new Minecraft blocks.
 
+### 1.5.2
+
+- Updated model injections to use Fabric API's new model loading API which should improve mod compatibility.
+- Added Vietnamese translations ([#90](https://github.com/LambdAurora/LambdaBetterGrass/pull/90)).
+
 [SpruceUI]: https://github.com/LambdAurora/SpruceUI
 [Connected Block Textures]: https://www.curseforge.com/minecraft/mc-mods/connected-block-textures "Connected Block Textures CurseForge page"